Remove unnecessary code from user functions & move to services/
