+++ src/services/users.js
@@ -1,0 +1,6 @@
+import axios from 'axios';
+
+const getUsers = () => axios.get('/users');
+const getUserInfo = username => axios.get(`/users/${username}`);
+
+export { getUsers, getUserInfo };
